api-connector/requirements.txt
coloredlogs==15.0.1
Flask==3.1.0
functions-framework==3.8.2
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.27.1
google-cloud-secret-manager==2.21.1
google-cloud-tasks==2.17.1
requests==2.32.3
